pyproject.toml
[project]
name = "automailsender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "bs4>=0.0.2",
    "pandas>=2.3.1",
    "pygithub>=2.7.0",
    "requests>=2.32.4",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
